Add files via upload
"This calendar is designed specifically for KIIT University students, offering an intuitive and user-friendly interface. It highlights key dates with distinctive colors, and when a user clicks on any of these dates, a pop-up provides detailed information about the event associated with that date. This calendar is tailored particularly for KIIT B.Tech students of the 2024 batch, covering all semesters."